+++ x/auth/types/account.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"strings"
 
+	gogoprotoany "github.com/cosmos/gogoproto/types/any"
+
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -14,11 +16,11 @@ import (
 )
 
 var (
-	_ sdk.AccountI                       = (*BaseAccount)(nil)
-	_ GenesisAccount                     = (*BaseAccount)(nil)
-	_ codectypes.UnpackInterfacesMessage = (*BaseAccount)(nil)
-	_ GenesisAccount                     = (*ModuleAccount)(nil)
-	_ sdk.ModuleAccountI                 = (*ModuleAccount)(nil)
+	_ sdk.AccountI                         = (*BaseAccount)(nil)
+	_ GenesisAccount                       = (*BaseAccount)(nil)
+	_ gogoprotoany.UnpackInterfacesMessage = (*BaseAccount)(nil)
+	_ GenesisAccount                       = (*ModuleAccount)(nil)
+	_ sdk.ModuleAccountI                   = (*ModuleAccount)(nil)
 )
 
 // NewBaseAccount creates a new BaseAccount object.
@@ -132,7 +134,7 @@ func (acc BaseAccount) Validate() error {
 }
 
 // UnpackInterfaces implements UnpackInterfacesMessage.UnpackInterfaces
-func (acc BaseAccount) UnpackInterfaces(unpacker codectypes.AnyUnpacker) error {
+func (acc BaseAccount) UnpackInterfaces(unpacker gogoprotoany.AnyUnpacker) error {
 	if acc.PubKey == nil {
 		return nil
 	}