+++ x/distribution/types/expected_keepers.go
@@ -34,14 +34,6 @@ type BankKeeper interface {
 	IsSendEnabledDenom(ctx context.Context, denom string) bool
 }
 
-// PoolKeeper defines the expected interface needed to fund & distribute pool balances.
-type PoolKeeper interface {
-	FundCommunityPool(ctx context.Context, amount sdk.Coins, sender sdk.AccAddress) error
-	DistributeFromCommunityPool(ctx context.Context, amount sdk.Coins, receiveAddr sdk.AccAddress) error
-	GetCommunityPool(ctx context.Context) (sdk.Coins, error)
-	SetToDistribute(ctx context.Context, amount sdk.Coins, addr string) error
-}
-
 // StakingKeeper expected staking keeper (noalias)
 type StakingKeeper interface {
 	ValidatorAddressCodec() address.Codec