+++ x/evidence/types/genesis_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 	"time"
 
+	gogoprotoany "github.com/cosmos/gogoproto/types/any"
 	"github.com/stretchr/testify/require"
 
 	"cosmossdk.io/x/evidence/exported"
@@ -130,7 +131,7 @@ func TestUnpackInterfaces(t *testing.T) {
 
 	testCases := []struct {
 		msg      string
-		unpacker codectypes.AnyUnpacker
+		unpacker gogoprotoany.AnyUnpacker
 		expPass  bool
 	}{
 		{